--- src/main/java/com/pnudev/springonlinedelivery/Role.java
@@ -1,5 +1,0 @@
-package com.pnudev.springonlinedelivery;
-
-public enum Role {
-	GUEST, USER, ADMIN
-}
