+++ src/main/java/com/pnudev/springonlinedelivery/models/Role.java
@@ -1,0 +1,5 @@
+package com.pnudev.springonlinedelivery.models;
+
+public enum Role {
+	GUEST, USER, ADMIN
+}
