+++ src/main/java/com/pnudev/springonlinedelivery/models/Status.java
@@ -1,4 +1,4 @@
-package com.pnudev.springonlinedelivery;
+package com.pnudev.springonlinedelivery.models;
 
 public enum Status {
 	ORDERED, DONE, CANCELED
